Rename Wire.h to WireStorage.h
